--- src/ensembl/src/content/app/browser/drawer/drawerState.ts
@@ -1,61 +1,0 @@
-/**
- * See the NOTICE file distributed with this work for additional information
- * regarding copyright ownership.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-type BookmarksDrawerView = {
-  name: 'bookmarks';
-};
-
-type GeneDrawerView = {
-  name: 'gene_summary';
-  payload: {
-    geneId: string; // in ensObjectId format
-  };
-};
-
-type TranscriptDrawerView = {
-  name: 'gene_summary';
-  payload: {
-    geneId: string; // in ensObjectId format
-    transcriptId: string; // transcript stable id
-  };
-};
-
-type GenericTrackView = {
-  name: 'track_details';
-  payload: {
-    trackId: string;
-  };
-};
-
-type DrawerView =
-  | BookmarksDrawerView
-  | GeneDrawerView
-  | TranscriptDrawerView
-  | GenericTrackView;
-
-export type DrawerStateForGenome = Readonly<{
-  isDrawerOpened: boolean;
-  drawerView: DrawerView | null;
-}>;
-
-export type DrawerState = Readonly<{
-  [genomeId: string]: DrawerStateForGenome;
-}>;
-
-export const defaultDrawerStateForGenome: DrawerStateForGenome = {
-  isDrawerOpened: false,
-  drawerView: null
-};
