+++ src/ensembl/src/objects.ts
@@ -19,7 +19,6 @@ import { RefObject, ReactEventHandler } from 'react';
 import * as headerActions from './header/headerActions';
 import * as browserActions from './content/app/browser/browserActions';
 import * as customDownloadActions from './content/app/custom-download/state/customDownloadActions';
-import * as drawerActions from './content/app/browser/drawer/drawerActions';
 import * as trackPanelActions from './content/app/browser/track-panel/trackPanelActions';
 
 export type ReactRefs = {
@@ -33,6 +32,5 @@ export type EventHandlers = {
 export type RootAction =
   | typeof headerActions
   | typeof browserActions
-  | typeof drawerActions
   | typeof customDownloadActions
   | typeof trackPanelActions;
